+++ stereotype/fields/annotations.py
@@ -10,7 +10,7 @@ if TYPE_CHECKING:  # pragma: no cover
 
 try:
     _py310_union_type = types.UnionType  # Python 3.10 or above
-except AttributeError:
+except AttributeError:  # pragma: no cover
     _py310_union_type = NotImplemented  # Python 3.9 or below
 
 
@@ -42,7 +42,9 @@ class AnnotationResolver:
         self.origin = get_origin(self.annotation)
 
     def __repr__(self):
-        return self.annotation.__name__ if hasattr(self.annotation, '__name__') else repr(self.annotation)
+        if self.origin is None and hasattr(self.annotation, '__name__'):
+            return self.annotation.__name__
+        return repr(self.annotation)
 
     def resolve(self, explict_field: Optional[Field] = None) -> Field:
         if explict_field is None or explict_field is NotImplemented: